final_project/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
blinker==1.5
cachelib==0.9.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cs50==9.2.2
DateTime==4.6
discord==2.1.0
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
flask-requests==0.0.14
Flask-Session==0.4.0
Flask-Sessions==0.1.5
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
helpers==0.2.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
platformdirs==2.5.2
protobuf==4.21.12
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sql==2022.4.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
termcolor==2.0.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
yarl==1.8.2
zope.interface==5.4.0
